removed duplicate index.ts file and updated links

+++ src/classes/User.ts
@@ -10,8 +10,8 @@ export class User {
     constructor () {
         this.name = faker.name.firstName();
         this.location = {
-            lat: parseInt(faker.address.latitude()),
-            lng: parseInt(faker.address.longitude())
+            lat: parseFloat(faker.address.latitude()),
+            lng: parseFloat(faker.address.longitude())
         };
 
     }
